+++ src/index.tsx
@@ -24,7 +24,10 @@ import { Logo } from "./components/Logo/Logo";
 import { Modal } from "./components/Modal/Modal";
 import { Radio } from "./components/Radio/Radio";
 import { RadioGroup } from "./components/Radio/RadioGroup";
+import { SiteAlert } from "./components/SiteAlert/SiteAlert";
 import { Search } from "./components/Search/Search";
+import { TabPanel } from "./components/Tabs/TabPanel";
+import { Tabs } from "./components/Tabs/Tabs";
 import { TextArea } from "./components/TextArea/TextArea";
 import { TextInput } from "./components/TextInput/TextInput";
 import { Typography } from "./components/Typography/Typography";
@@ -53,6 +56,9 @@ export {
   Radio,
   RadioGroup,
   Search,
+  SiteAlert,
+  TabPanel,
+  Tabs,
   TextArea,
   TextInput,
   Typography,